feat: one command runs everything, and fix bars that never filled
`npm start` launches the bot and the site together. The alternative was two
terminals, two commands and a remembered order, and the person running this
should not have to hold that. If either child dies the other is stopped: a
site serving a database nobody is updating looks healthy while being hours
stale, and being plainly down is more honest than that. On an empty
database it refuses to start and prints the one command that fixes it,
rather than showing a page of empty panels whose reason is three screens up
in a log.

Fixes a CSS bug found by actually looking at the rendered page: the funnel
and confidence bars are `<span>` elements, a span is inline by default, and
an inline element IGNORES width and height. Every bar had been rendering as
an empty track since it was written. They fill now.

And the demo seed no longer includes short signals. It was showing the
operator a kind of signal the live system — spot only — can never produce,
and a demonstration of the wrong product is worse than none: it teaches the
reader to expect something that cannot arrive.

Verified by loading all eight pages in a real browser and reading them.

+++ scripts/seed-demo.ts
@@ -11,6 +11,11 @@
  *
  * Writes to data/demo.db unless DB_FILE says otherwise, so it cannot
  * overwrite a real run.
+ *
+ * Every seeded signal is a BUY. The demo used to include shorts, which showed
+ * the operator a kind of signal the live system — spot only — will never
+ * produce. A demonstration that demonstrates the wrong product is worse than
+ * none: it teaches the reader to expect something that cannot arrive.
  */
 import crypto from "node:crypto";
 import { openDb, closeDb } from "@/storage/db";
@@ -172,11 +177,11 @@ function main(): void {
   }[] = [
     { symbol: "SOLUSDT", tf: "4h", direction: "long", setup: "trend_continuation", confidence: 74, outcome: "win2", ageHours: 300 },
     { symbol: "ETHUSDT", tf: "1h", direction: "long", setup: "breakout_retest", confidence: 68, outcome: "win1", ageHours: 240 },
-    { symbol: "LINKUSDT", tf: "4h", direction: "short", setup: "range_reversal", confidence: 55, outcome: "loss", ageHours: 200 },
+    { symbol: "LINKUSDT", tf: "4h", direction: "long", setup: "range_reversal", confidence: 55, outcome: "loss", ageHours: 200 },
     { symbol: "AVAXUSDT", tf: "1h", direction: "long", setup: "momentum_ignition", confidence: 71, outcome: "win2", ageHours: 160 },
     { symbol: "ADAUSDT", tf: "1d", direction: "long", setup: "trend_continuation", confidence: 49, outcome: "loss", ageHours: 130 },
     { symbol: "BTCUSDT", tf: "4h", direction: "long", setup: "liquidity_sweep", confidence: 66, outcome: "win1", ageHours: 96 },
-    { symbol: "ETHUSDT", tf: "4h", direction: "short", setup: "divergence_reversal", confidence: 52, outcome: "loss", ageHours: 72 },
+    { symbol: "ETHUSDT", tf: "4h", direction: "long", setup: "divergence_reversal", confidence: 52, outcome: "loss", ageHours: 72 },
     { symbol: "SOLUSDT", tf: "1h", direction: "long", setup: "trend_continuation", confidence: 63, outcome: "open", ageHours: 18 },
     { symbol: "BTCUSDT", tf: "1h", direction: "long", setup: "breakout_retest", confidence: 58, outcome: "pending", ageHours: 3 },
   ];
